backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
paho-mqtt==1.6.1
python-dotenv==1.0.0
psycopg2-binary==2.9.10
gunicorn==22.0.0